Spyglass Tagged object highlights
So, it's looking like the idea I originally had requires duplicating the tagged object and childing the duplicate under the original, scaling it to be very slightly larger by percentage or pixels and the applying the overlay material with a mostly translucent setting, and then destroying the duplicate when spyglass is no longer active.  What's in here is essentially a copy paste of what I did fo the environment material switch.  It's a lot simpler and functional, if we end up not having time or not caring enough to do the crazy thing.